--- src/QuickMarcEditor/QuickMarcEditorRows/Leader/index.js
@@ -1,1 +1,0 @@
-export * from './Leader';
