+++ src/QuickMarcEditor/QuickMarcEditorRows/LeaderField/leaderConfig.js
